Refactored packages and added tar.gz reader
