+++ doc/notes/zrq/20190904-01-kafka-tests.txt
@@ -78,6 +78,7 @@
 
     rewind=true
     rewind=false
+    rewind=true
 
     pushd "${PHYMATOPUS_CODE:?}"
         pushd kafka-tools
@@ -86,9 +87,13 @@
 
             mvn clean compile
 
+            topic=ztf_20190916_programid1            
             topic=ztf_20190904_programid1
+
+            topic=ztf_20190823_programid1
             looplimit=10000
-            threads=4
+            threads=8
+            rewind=true
 
             mvn test \
                 -D 'test=ZtfAlertReaderTest' \
