+++ doc/notes/zrq/20200322-03-design-ideas.txt
@@ -1,0 +1,110 @@
+#
+# <meta:header>
+#   <meta:licence>
+#     Copyright (c) 2020, ROE (http://www.roe.ac.uk/)
+#
+#     This information is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
+#
+#     This information is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#     GNU General Public License for more details.
+#
+#     You should have received a copy of the GNU General Public License
+#     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#   </meta:licence>
+# </meta:header>
+#
+#zrq-notes-time
+#zrq-notes-indent
+#zrq-notes-crypto
+#zrq-notes-ansible
+#zrq-notes-osformat
+#
+
+# -----------------------------------------------------
+# Components we could build ..
+
+    # Read Kafka, store images -> blob.
+    # Read Kafka, store curves -> blob.
+    # Read Kafka, store curves -> Cassandra.
+    
+    # Read messages, calculate apparent magnitude -> Kafka 
+    # Read messages, calculate curve features -> Kafka 
+    # Read messages, calculate htmid -> Kafka 
+   
+    # Output logging statistics to a separtate Kafka stream
+    # Replay the logging stream to terminal
+
+    # Relying on stored data on trop03 is fragile.
+    # Can we load the buffer from tar.gz Avro files ?
+    # Read Avro tar.gz output to Kafka.
+
+    Example Kafka deploy - trop-03, trop-04
+        Refactor to spread the disc load
+        
+    Example Cassandra deploy - gaia-01, gaia-02 ?
+        /dev/nvme0n1p6  372G  5.5G  347G   2% /var
+        /dev/nvme0n1p6  372G  769M  352G   1% /var
+
+    Trop nodes
+        32 core
+        128G RAM
+        1 x 240GB SSD
+        2 x   4T HDD
+
+            /dev/sda2 92G /
+            /dev/sda4 10G /tmp
+            /dev/sda5 65G /var
+            /dev/sda6 53G /home
+
+    Worker nodes
+        28 core
+        512G RAM
+        1 x 480GB SSD
+        3 x   8TB HDD
+
+            /dev/sda1  92G /
+            /dev/sda5  37G /tmp
+            /dev/sda6 304G /var
+
+    Gaia Nodes
+        32 core
+        256G RAM
+        1 x 512GB NVMe
+        4 x   8TB HDD
+
+            /dev/nvme0n1p1   46G /
+            /dev/nvme0n1p6  372G /var
+            /dev/nvme0n1p5   37G /tmp
+        
+# -----------------------------------------------------
+# -----------------------------------------------------
+# Read a ZTF tar.gz and extract Avro ...
+#[user@desktop]
+
+    source "${HOME:?}/phymatopus.env"
+    pushd "${PHYMATOPUS_CODE}"
+
+        mkdir data    
+        pushd data    
+
+            mkdir archive
+            pushd archive
+
+                wget https://ztf.uw.edu/alerts/public/ztf_public_20191206.tar.gz
+
+--START--
+--2020-03-22 17:30:24--  https://ztf.uw.edu/alerts/public/ztf_public_20191206.tar.gz
+....
+2020-03-22 17:47:56 (1.58 MB/s) - ‘ztf_public_20191206.tar.gz’ saved [1739933850/1739933850]
+--END--
+
+
+
+
+
+
