+++ kafka-tools/pom.xml
@@ -296,6 +296,15 @@
         </dependency>
 
         <!--+
+            | Apache Commons Compress.
+            +-->
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-compress</artifactId>
+            <version>1.20</version>
+        </dependency>
+
+        <!--+
             | JUnit test framework.
             | http://www.junit.org/
             +-->
