+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/alert/AlertProcessor.java
@@ -15,20 +15,25 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package uk.ac.roe.wfau.phymatopus.kafka.alert.ztf;
 
-public interface ZtfCutout
+package uk.ac.roe.wfau.phymatopus.alert;
+
+/**
+ * Alert processor interface.
+ * 
+ */
+public interface AlertProcessor<AlertType>
     {
     /**
-     * Gets the value of the 'fileName' field.
-     * @return The value of the 'fileName' field.
+     * Process an alert.
+     * 
      */
-    public java.lang.CharSequence getFileName();
+    public void process(final AlertType alert);
 
     /**
-     * Gets the value of the 'stampData' field.
-     * @return fits.gz
+     * The number of alerts this processor has handled. 
+     * 
      */
-    public java.nio.ByteBuffer getStampData();
+    public long count();
 
     }
