+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/alert/BaseCandidate.java
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package uk.ac.roe.wfau.phymatopus.kafka.alert;
+package uk.ac.roe.wfau.phymatopus.alert;
 
 public interface BaseCandidate
     {
