+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/alert/PrevCandidate.java
@@ -15,36 +15,14 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package uk.ac.roe.wfau.phymatopus.kafka.alert.ztf;
+package uk.ac.roe.wfau.phymatopus.alert;
 
-import java.nio.ByteBuffer;
-
-public class ZtfCutoutWrapper implements ZtfCutout
+/**
+ * 
+ *
+ */
+public interface PrevCandidate
+extends BaseCandidate
     {
-    /**
-     * Our Avro alert bean.
-     *
-     */
-    private ztf.cutout bean ;
 
-    /**
-     * Public constructor.
-     * 
-     */
-    public ZtfCutoutWrapper(final ztf.cutout bean)
-        {
-        this.bean = bean;
-        }
-
-    @Override
-    public CharSequence getFileName()
-        {
-        return bean.getFileName();
-        }
-
-    @Override
-    public ByteBuffer getStampData()
-        {
-        return bean.getStampData();
-        }
     }
