+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/avro/lsst/LsstAlertCandidateWrapper.java
@@ -15,12 +15,12 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package uk.ac.roe.wfau.phymatopus.kafka.alert.lsst;
+package uk.ac.roe.wfau.phymatopus.avro.lsst;
 
 import org.apache.avro.generic.GenericData;
 
 import lombok.extern.slf4j.Slf4j;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.AlertCandidate;
+import uk.ac.roe.wfau.phymatopus.alert.AlertCandidate;
 
 @Slf4j
 public class LsstAlertCandidateWrapper
