+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/avro/ztf/ZtfCutoutWrapper.java
@@ -15,14 +15,36 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-package uk.ac.roe.wfau.phymatopus.kafka.alert;
+package uk.ac.roe.wfau.phymatopus.avro.ztf;
 
-/**
- * 
- *
- */
-public interface PrevCandidate
-extends BaseCandidate
+import java.nio.ByteBuffer;
+
+public class ZtfCutoutWrapper implements ZtfCutout
     {
+    /**
+     * Our Avro alert bean.
+     *
+     */
+    private ztf.cutout bean ;
 
+    /**
+     * Public constructor.
+     * 
+     */
+    public ZtfCutoutWrapper(final ztf.cutout bean)
+        {
+        this.bean = bean;
+        }
+
+    @Override
+    public CharSequence getFileName()
+        {
+        return bean.getFileName();
+        }
+
+    @Override
+    public ByteBuffer getStampData()
+        {
+        return bean.getStampData();
+        }
     }
