+++ kafka-tools/src/main/java/uk/ac/roe/wfau/phymatopus/kafka/alert/GenericAlertReader.java
@@ -35,6 +35,8 @@ import org.apache.kafka.common.PartitionInfo;
 import org.apache.kafka.common.TopicPartition;
 
 import lombok.extern.slf4j.Slf4j;
+import uk.ac.roe.wfau.phymatopus.alert.AlertProcessor;
+import uk.ac.roe.wfau.phymatopus.alert.AlertReader;
 import uk.ac.roe.wfau.phymatopus.kafka.BaseReader;
 
 /**
