+++ kafka-tools/src/test/conf/logback.xml
@@ -45,7 +45,7 @@
 	    | Phymatopus logging.
 	    +-->
 	<logger name="uk.ac.roe.wfau.phymatopus">
-		<level value="INFO"/>
+		<level value="DEBUG"/>
 	</logger>
 
 	<!--+
