+++ kafka-tools/src/test/java/uk/ac/roe/wfau/phymatopus/cassandra/CassandraWriterTestBase.java
@@ -21,10 +21,10 @@ package uk.ac.roe.wfau.phymatopus.cassandra;
 import org.springframework.beans.factory.annotation.Value;
 
 import lombok.extern.slf4j.Slf4j;
+import uk.ac.roe.wfau.phymatopus.alert.AlertProcessor;
+import uk.ac.roe.wfau.phymatopus.alert.BaseAlert;
 import uk.ac.roe.wfau.phymatopus.cassandra.AbstractCassandraWriter;
 import uk.ac.roe.wfau.phymatopus.kafka.KafkaReaderTestBase;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.AlertProcessor;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.BaseAlert;
 
 /**
  *
