+++ kafka-tools/src/test/java/uk/ac/roe/wfau/phymatopus/cassandra/ZtfCandiateWriterTest.java
@@ -24,9 +24,9 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import lombok.extern.slf4j.Slf4j;
+import uk.ac.roe.wfau.phymatopus.alert.AlertReader;
 import uk.ac.roe.wfau.phymatopus.cassandra.AbstractCassandraWriter;
 import uk.ac.roe.wfau.phymatopus.cassandra.CandiateWriter;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.AlertReader;
 import uk.ac.roe.wfau.phymatopus.kafka.alert.ztf.ZtfAlertReader;
 
 /**
