+++ kafka-tools/src/test/java/uk/ac/roe/wfau/phymatopus/kafka/tools/StringWriterTest.java
@@ -21,10 +21,10 @@ package uk.ac.roe.wfau.phymatopus.kafka.tools;
 import org.junit.Test;
 
 import lombok.extern.slf4j.Slf4j;
+import uk.ac.roe.wfau.phymatopus.alert.AlertProcessor;
+import uk.ac.roe.wfau.phymatopus.alert.AlertReader.CallableAlertReader;
 import uk.ac.roe.wfau.phymatopus.kafka.BaseClient;
 import uk.ac.roe.wfau.phymatopus.kafka.KafkaReaderTestBase;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.AlertProcessor;
-import uk.ac.roe.wfau.phymatopus.kafka.alert.AlertReader.CallableAlertReader;
 
 /**
  *
